backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-admirarchy==1.2.2
django-choices==1.7.2
django-environ==0.10.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-recurrence==1.11.1
django-rest-swagger==2.2.0
django-scheduler==0.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
icalendar==5.0.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
views==0.3
